Move Mongo vCore vectorstore to `langchain-azure-cosmosdb` and add deprecation shims in `langchain-azure-ai` (#870)
This pull request primarily renames and updates references from
`AzureCosmosDBMongoVCoreVectorSearch` to `AzureDocumentDBVectorSearch`
throughout the codebase, reflecting a migration to a new or rebranded
vector search implementation. It also updates imports to use the new
module, ensures compatibility in tests, and updates type annotations
accordingly.

Key changes include:

**Refactoring and Renaming:**
* Replaced all usages of `AzureCosmosDBMongoVCoreVectorSearch` with
`AzureDocumentDBVectorSearch` in both implementation and test files,
including class names, function calls, and type annotations.
[[1]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L1-R17)
[[2]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L74-R77)
[[3]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L109-R112)
[[4]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L149-R152)
[[5]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L186-R189)
[[6]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L221-R224)
[[7]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L259-R262)
[[8]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L312-R315)
[[9]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L371-R374)
[[10]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L409-R412)
[[11]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L441-R444)
[[12]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L474-R477)
[[13]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L518-R521)
[[14]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L562-R565)
[[15]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L604-R607)
[[16]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L644-R647)
[[17]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L687-R690)
[[18]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L746-R749)
[[19]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L810-R813)
[[20]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L847-R850)
[[21]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L885-R888)
[[22]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L934-R937)
[[23]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L976-R979)
[[24]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L1016-R1019)
[[25]](diffhunk://#diff-ddad57423de1acd0679386cdfdd3a2312434bd6ef770c2207bec4eff4a300471L256-R255)
[[26]](diffhunk://#diff-ddad57423de1acd0679386cdfdd3a2312434bd6ef770c2207bec4eff4a300471L267-R266)
[[27]](diffhunk://#diff-ddad57423de1acd0679386cdfdd3a2312434bd6ef770c2207bec4eff4a300471L279-R285)

**Import and Module Adjustments:**
* Updated imports in `vectorstores/__init__.py` and `cache.py` to
reference `AzureDocumentDBVectorSearch` and its related types from the
new `langchain_azure_cosmosdb` package.
[[1]](diffhunk://#diff-29f824e37c547b6d6743d8ae2eb9944dcf91c350aee43f3f45c22734ea07810aL25-R37)
[[2]](diffhunk://#diff-ddad57423de1acd0679386cdfdd3a2312434bd6ef770c2207bec4eff4a300471R24-L36)

**Test Suite Compatibility:**
* Modified integration tests to import and test
`AzureDocumentDBVectorSearch` instead of the previous class, and added
`pytest.importorskip("langchain_azure_cosmosdb")` to skip tests if the
new dependency is unavailable.
[[1]](diffhunk://#diff-9f5f5133a4abfdd4b00bbfc87beb97a5d9339e12e845ce653b283a07d0580d10L1-R17)
[[2]](diffhunk://#diff-bd1969f174eace39e0457fa2580485602b97bf0fba4433aaa35a6a7c3570d028R16-R18)

These changes collectively update the codebase to use the new vector
search implementation, ensure type correctness, and maintain test
coverage with the updated class.

---------
