+++ libs/azure-cosmosdb/tests/unit_tests/test_imports.py
@@ -7,9 +7,13 @@ EXPECTED_ALL = [
     "AsyncCosmosDBChatMessageHistory",
     "AsyncCosmosDBStore",
     "AzureCosmosDBNoSqlSemanticCache",
+    "AzureDocumentDBVectorSearch",
     "AzureCosmosDBNoSqlVectorSearch",
     "AzureCosmosDBNoSqlVectorStoreRetriever",
     "AzureCosmosDbNoSQLTranslator",
+    "CosmosDBSimilarityType",
+    "CosmosDBVectorSearchCompression",
+    "CosmosDBVectorSearchType",
     "CosmosDBCache",
     "CosmosDBCacheSync",
     "CosmosDBChatMessageHistory",
